move get_hf_config_content to lmdeploy.utils
